+++ Libft/ft_memchr.c
@@ -1,6 +1,8 @@
-void	*ft_memchr(const void *s, int c, unsigned int n)
+#include "libft.h"
+
+void	*ft_memchr(const void *s, int c, size_t n)
 {
-	unsigned int	i;
+	size_t	i;
 	int	c_found;
 
 	i = 0;
