+++ Libft/ft_memcmp.c
@@ -1,6 +1,8 @@
-int	ft_memcmp(const void *s1, const void *s2, unsigned int n)
+#include "libft.h"
+
+int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n && ((unsigned char *)s1)[i] == ((unsigned char *)s2)[i])
