+++ Libft/ft_memcpy.c
@@ -1,4 +1,6 @@
-void	*memcpy(void *dest, const void *src, unsigned int n)
+#include "libft.h"
+
+void	*memcpy(void *dest, const void *src, size_t n)
 {
 	n--;
 	while (n > 0)
